refactor: refresh dependencies, competition code improvement
