+++ build.gradle
@@ -60,5 +60,5 @@ task clean(type: Delete) {
 ext {
     roomVersion = '2.2.5'
     coroutines = '1.3.7'
-    daggerVersion = '2.27'
+    daggerVersion = '2.28'
 }
